Half-ellipse fixing
I did not want all of the half-ellipses to each be their own shape, as it would increase the chance that the system would pick more half-ellipses instead of the other shapes.

My solution was to make one shape called "half-ellipse". Then I added if-statements with their own unique rotation of the half-ellipse. When the system chooses half-ellipse as the shape, the if-statement will decide the rotation.